clean up

+++ src/basic_types/mod.rs
@@ -23,7 +23,7 @@ basic_type!(
 basic_type!(
     /// Unique identifier of the account in the Fluidex network.
     AccountId,
-    u32 // TODO: u16?
+    u16 // TODO: u32?
 );
 
 basic_type!(
